requirements.txt
cx-Oracle==7.2.2
mysql-connector-python==8.0.17
protobuf==3.9.1
psycopg2==2.8.3
PyMySQL==0.9.3
pyodbc==4.0.27
six==1.12.0

